Rename historic missions to mission history
