+++ frontend/src/components/Pages/MissionHistoryPage/HistoricMissionCard.tsx
